feat: add imput robot and state robot
