+++ index.js
@@ -1,31 +1,16 @@
-const readline = require('readline-sync')
 const robots = {
-    text: require('./robots/text.js')
+    imput: require('./robots/imput.js'),
+    text: require('./robots/text.js'),
+    state: require('./robots/state.js')
 }
 
 async function start() {
-    const content = {
-        maximumSentences: 7
-    }
+    robots.imput()
+    await robots.text()
 
-    content.searchTerm = askAndReturnSearchTerm();
-    content.prefix = askAndReturnPrefix();
+    const content = robots.state.load()
+    console.dir(content, { depth: null})
 
-    await robots.text(content)
-
-    function askAndReturnSearchTerm() {
-        return readline.question('Type a Wikipedia search term: ');
-    }
-
-    function askAndReturnPrefix() {
-        const prefixes = ['Who is', 'What is', 'The history of'];
-        const selectedPrefixIndex = readline.keyInSelect(prefixes, 'Choose one option: ');
-        const selectedPrefixText = prefixes[selectedPrefixIndex];
-
-        return selectedPrefixText;
-    }
-
-    console.log(JSON.stringify(content, null, 4));
 }
 
 start();
